requirements.txt
# Core dependencies
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
protobuf==4.25.1

# Text processing
nltk==3.8.1
spacy==3.7.2

# Language detection
langdetect==1.0.9

# Utilities
numpy==1.24.3
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CLI interface
click==8.1.7

# File handling
python-docx==1.1.0
PyPDF2==3.0.1

